requirements.txt
bson==0.5.10
click==8.0.0
Flask==2.0.0
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
numpy==1.20.3
pandas==1.2.4
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
sklearn==0.0
threadpoolctl==2.1.0
Werkzeug==2.0.1
